Moved rect struct into its own folder
