+++ src/plater/placer.rs
@@ -1,4 +1,3 @@
-use std::cmp::Ordering;
 use std::collections::HashMap;
 use std::error::Error;
 use std::fmt::Debug;
@@ -9,6 +8,7 @@ use rand::thread_rng;
 
 use crate::plater::placed_part::PlacedPart;
 use crate::plater::placer::helpers::find_solution;
+use crate::plater::placer::rect::Rect;
 use crate::plater::placer::score::ScoreOrder;
 use crate::plater::placer::search::exponential_search;
 use crate::plater::placer::GravityMode::{GravityEQ, GravityXY, GravityYX};
@@ -18,60 +18,6 @@ use crate::plater::request::{BedExpansionMode, Request};
 use crate::plater::solution::Solution;
 
 pub(crate) const N: usize = 1024;
-
-#[derive(Clone, Copy, Debug, PartialEq)]
-pub struct Rect {
-    width: f64,
-    height: f64,
-    center_x: f64,
-    center_y: f64,
-}
-
-impl Rect {
-    fn combine(&self, other: &Self) -> Self {
-        let top_height = f64::max(
-            self.height / 2.0 + self.center_y,
-            other.height / 2.0 + other.center_y,
-        );
-        let bottom_height = f64::min(
-            -self.height / 2.0 + self.center_y,
-            -other.height / 2.0 + other.center_y,
-        );
-
-        let left_width = f64::min(
-            -self.width / 2.0 + self.center_x,
-            -other.width / 2.0 + other.center_x,
-        );
-        let right_width = f64::max(
-            self.width / 2.0 + self.center_x,
-            other.width / 2.0 + other.center_x,
-        );
-
-        Rect {
-            width: right_width - left_width,
-            height: top_height - bottom_height,
-            center_x: (right_width + left_width) / 2.0,
-            center_y: (top_height + bottom_height) / 2.0,
-        }
-    }
-
-    fn get_points(&self) -> [(f64, f64); 4] {
-        let w2 = self.width / 2.0;
-        let h2 = self.height / 2.0;
-        [
-            (self.center_x + w2, self.center_y + h2),
-            (self.center_x - w2, self.center_y + h2),
-            (self.center_x + w2, self.center_y - h2),
-            (self.center_x - w2, self.center_y - h2),
-        ]
-    }
-}
-
-impl PartialOrd for Rect {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        f64::partial_cmp(&(self.width * self.height), &(other.width * other.height))
-    }
-}
 
 #[derive(Clone, Copy)]
 pub enum SortMode {
@@ -438,6 +384,7 @@ pub(crate) fn all_parts_can_eventually_be_attempted(
 }
 
 mod helpers;
+mod rect;
 pub(crate) mod score;
 mod search;
 mod strategies;
